fix is_valid of PoolFile to compare a long with a long, not a str

+++ daklib/dbconn.py
@@ -1162,7 +1162,7 @@ class PoolFile(ORMObject):
         return os.path.join(self.location.path, self.filename)
 
     def is_valid(self, filesize = -1, md5sum = None):\
-        return self.filesize == filesize and self.md5sum == md5sum
+        return self.filesize == long(filesize) and self.md5sum == md5sum
 
     def properties(self):
         return ['filename', 'file_id', 'filesize', 'md5sum', 'sha1sum', \
